feat(innexus): add waveguide string resonator with modal/waveguide switching
Implement digital waveguide string resonator (Karplus-Strong/EKS) as an
alternative to the existing modal resonator bank in Innexus. Users can
switch between Modal and Waveguide resonance types with click-free
equal-power cosine crossfade and energy-aware gain matching.

DSP core (dsp/include/krate/dsp/processors/):
- IResonator: shared pure virtual interface for interchangeable resonators
- WaveguideString: two-segment delay loop with loss filter, dispersion
  allpass cascade (4 sections), DC blocker, soft clipper, energy floor,
  dual energy followers, noise burst excitation with pick-position comb
- ModalResonatorBank: adapted to conform to IResonator interface

Plugin integration (plugins/innexus/):
- Three new parameters: ResonanceType, WaveguideStiffness, WaveguidePickPosition
- Voice engine wiring with per-voice WaveguideString instances
- 1024-sample crossfade with equal-power formula and energy gain matching
- Full state save/load and controller registration

Key features:
- Sub-1-cent pitch accuracy (A2-A5) via exact phase delay compensation
- Stiffness control for inharmonicity shaping (Fletcher's formula)
- Pick position comb filter with circular subtraction for deep spectral nulls
- Energy normalisation for consistent loudness across pitch range (within 3 dB)
- ScatteringJunction/PluckJunction stubs for future bow model (Phase 4)

Also fixes 63 clang-tidy warnings across disrumpo, ruinae, and innexus plugins.

Closes #129

+++ dsp/tests/unit/core/hungarian_algorithm_test.cpp
@@ -158,7 +158,7 @@ TEST_CASE("HungarianAlgorithm: frequency distance cost matrix for partial tracki
     for (int i = 0; i < 3; ++i) {
         for (int j = 0; j < 4; ++j) {
             float dist = std::abs(prevFreqs[i] - peakFreqs[j]);
-            cost[static_cast<size_t>(i * 4 + j)] = (dist < kMaxDist) ? dist : kMaxDist * 100.0f;
+            cost[static_cast<size_t>(i) * 4 + static_cast<size_t>(j)] = (dist < kMaxDist) ? dist : kMaxDist * 100.0f;
         }
     }
 
@@ -183,7 +183,7 @@ TEST_CASE("HungarianAlgorithm: crossing partials resolved correctly",
     std::array<float, 4> cost;
     for (int i = 0; i < 2; ++i) {
         for (int j = 0; j < 2; ++j) {
-            cost[static_cast<size_t>(i * 2 + j)] = std::abs(prevFreqs[i] - peakFreqs[j]);
+            cost[static_cast<size_t>(i) * 2 + static_cast<size_t>(j)] = std::abs(prevFreqs[i] - peakFreqs[j]);
         }
     }
     // cost = [8, 2, 2, 8]
