+++ plugins/disrumpo/src/plugin_ids.h
@@ -690,7 +690,7 @@ constexpr int32_t kPresetVersion = 12;
 /// Migration table: old note index (v9, 15 entries) -> new dropdown index (v10, 21 entries).
 /// Old encoding: NoteValue * 3 + NoteModifier (Whole→Sixteenth, None/Dotted/Triplet).
 /// New encoding: kNoteValueDropdownMapping[] index (1/64T→4/1D).
-constexpr int8_t kOldNoteIdxToNewDropdown[15] = {
+constexpr int kOldNoteIdxToNewDropdown[15] = {
     19, 20, 18,  // 1/1, 1/1D, 1/1T
     16, 17, 15,  // 1/2, 1/2D, 1/2T
     13, 14, 12,  // 1/4, 1/4D, 1/4T
